GVNEditor/requirements.txt
PyQt5==5.15.2
PyYAML==5.4.1
